refactor: move images sign package
