+++ handlers/blogs/blogsPage.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/arran4/goa4web/a4code/a4code2html"
 	"github.com/arran4/goa4web/core"
-	imagesign "github.com/arran4/goa4web/pkg/images"
+	imagesign "github.com/arran4/goa4web/internal/images"
 	"github.com/gorilla/feeds"
 )
