+++ handlers/forum/forumFeed.go
@@ -14,7 +14,7 @@ import (
 	"github.com/arran4/goa4web/core"
 	common "github.com/arran4/goa4web/core/common"
 	db "github.com/arran4/goa4web/internal/db"
-	imagesign "github.com/arran4/goa4web/pkg/images"
+	imagesign "github.com/arran4/goa4web/internal/images"
 	"github.com/gorilla/feeds"
 	"github.com/gorilla/mux"
 )
