+++ handlers/imagebbs/imagebbsFeed.go
@@ -15,7 +15,7 @@ import (
 	common "github.com/arran4/goa4web/core/common"
 	"github.com/arran4/goa4web/core/templates"
 	db "github.com/arran4/goa4web/internal/db"
-	imagesign "github.com/arran4/goa4web/pkg/images"
+	imagesign "github.com/arran4/goa4web/internal/images"
 	"github.com/gorilla/feeds"
 	"github.com/gorilla/mux"
 )
