+++ handlers/images/upload_task.go
@@ -15,9 +15,9 @@ import (
 	"github.com/arran4/goa4web/config"
 	common "github.com/arran4/goa4web/core/common"
 	db "github.com/arran4/goa4web/internal/db"
+	imagesign "github.com/arran4/goa4web/internal/images"
 	"github.com/arran4/goa4web/internal/tasks"
 	"github.com/arran4/goa4web/internal/upload"
-	imagesign "github.com/arran4/goa4web/pkg/images"
 	"github.com/disintegration/imaging"
 )
