+++ handlers/linker/linkerFeed.go
@@ -12,7 +12,7 @@ import (
 	"github.com/arran4/goa4web/a4code/a4code2html"
 	common "github.com/arran4/goa4web/core/common"
 	db "github.com/arran4/goa4web/internal/db"
-	imagesign "github.com/arran4/goa4web/pkg/images"
+	imagesign "github.com/arran4/goa4web/internal/images"
 	"github.com/gorilla/feeds"
 )
