+++ handlers/user/userGalleryPage.go
@@ -14,7 +14,7 @@ import (
 	"github.com/arran4/goa4web/core"
 	handlers "github.com/arran4/goa4web/handlers"
 	db "github.com/arran4/goa4web/internal/db"
-	imagesign "github.com/arran4/goa4web/pkg/images"
+	imagesign "github.com/arran4/goa4web/internal/images"
 )
 
 type galleryImage struct {
