+++ handlers/writings/writingsFeed.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/arran4/goa4web/a4code/a4code2html"
 	common "github.com/arran4/goa4web/core/common"
-	imagesign "github.com/arran4/goa4web/pkg/images"
+	imagesign "github.com/arran4/goa4web/internal/images"
 	"github.com/gorilla/feeds"
 )
